backend/requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-storages
drf-spectacular
dj-database-url
psycopg2-binary
pillow
